Update README.md
A small program for keeping records and performing statistical analyses for a class of 20 students. Information about each student includes: ID, name, test results (3 tests per semester) and final score (average of tests). Use a class for this purpose - implement appropriate functionalities that can be used later.

In the main() function, the program will ask the user to select from the menu:
1. Add entry.
2. Delete entry.
3. Update entry.
4. Display all entries.
5. Calculate the average of the results of the selected student.
6. Display the average of the results of the selected student.
7. Calculate and display the average of all students.
8. Display the best and worst results (along with the data of the students who achieved them).